fix issues with import and export systems
